Improved xsl and fixed some issues
The tag '<font>' is not supported in HTML5. To add style, should be used CSS instead. This can be done with the attribute 'style'.
Added '<tbody>' tag to the table.
Edited the table style.
Made the table responsive. To achieve this the table has been moved inside a '<div>'.
Moved the title to '<head>'.