Fixed how are handled the questions in 'GestionPreguntas.php'
First the fields are passed by POST to the server to a php script.
The php checks that all the fields are valid.
Once is valid the values, they are added to the databese and the xml file.
When AJAX gets the values parsed into JSON format, refreshes the page showing a success message or a alert message to the user.
If all is correct the new xml will be shown, along with a success message.